requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
boto==2.46.1
bz2file==0.98
click==6.7
decorator==4.0.11
emoji==0.4.5
enum34==1.1.6
gensim==0.12.2
ipython==5.3.0
ipython-genutils==0.2.0
mecab-python==0.996
numpy==1.10.4
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
requests==2.14.2
scandir==1.5
scipy==0.16.1
simplegeneric==0.8.1
six==1.10.0
smart-open==1.5.2
tqdm==4.11.2
traitlets==4.3.2
wcwidth==0.1.7
zenhan==0.5.2
